ai_resume_analyzer/requirements.txt
Flask
pdfplumber
python-docx
openai  # or use requests for Ollama
